Move bus interaction handler.
