+++ constraint-solver/src/lib.rs
@@ -1,6 +1,7 @@
 //! Tooling used for analysis and solving of constraints.
 
 pub mod algebraic_constraint;
+pub mod bus_interaction_handler;
 pub mod constraint_system;
 pub mod effect;
 pub mod grouped_expression;
